+++ tests/testthat/test-plot_compact_pvd.R
@@ -11,6 +11,8 @@ test_that("`plot_compact_pvd()` produces consistent results", {
     )
    fig1 <- fig0 |>
     plot_compact_pvd(scale_colors = c("red", "blue", "purple4"),
+                     legend.position = "none",
+                     rel_heights = c(1, 0.1),
                      facet_label_prefix = NULL)
 
   fig1 |> vdiffr::expect_doppelganger(title = "pvd")
@@ -30,10 +32,32 @@ test_that(
 
     fig2 <-
       pvd_list_by_sex |>
-      plot_compact_pvd()
+      plot_compact_pvd(
+        legend.position = "none",
+        rel_heights = c(1, 0.1)
+      )
 
     fig2 |> vdiffr::expect_doppelganger(title = "pvd-fxtas-by-sex")
 
+    fig3 <-
+      pvd_list_by_sex |>
+      plot_compact_pvd(
+        legend.position = "right",
+        legend.direction = "vertical"
+      )
+
+    fig3 |> vdiffr::expect_doppelganger(title = "pvd-fxtas-by-sex-with-legend")
+
+    fig4 <-
+      pvd_list_by_sex |>
+      plot_compact_pvd(
+        legend.position = "right",
+        legend.direction = "vertical",
+        colorbar_label_type = "subscript"
+
+      )
+
+    fig4 |> vdiffr::expect_doppelganger(title = "pvd-fxtas-by-sex-with-legend-subscript")
 
   }
 )